ai/performance-predictor/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Machine Learning
torch==2.1.2
transformers==4.37.0

# External APIs
openai==1.10.0
httpx==0.26.0

# Caching
redis==5.0.1

# Database (for historical learning)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
joblib==1.3.2

# Image/Video Processing
pillow==10.2.0
opencv-python==4.9.0.80
